+++ packages/yourloops/components/patient-list/patient-list.hook.tsx
@@ -58,7 +58,7 @@ interface PatientListHookReturns {
   selectedTab: PatientListTabs
   inputSearch: string
   gridApiRef: MutableRefObject<GridApiCommon>
-  numberOfPatientsDisplayed: number
+  patientsDisplayedCount: number
   patientToRemoveForHcp: Patient | null
   patientToRemoveForCaregiver: UserToRemove | null
   rowsProps: GridRowsProp
@@ -260,7 +260,7 @@ export const usePatientListHook = (): PatientListHookReturns => {
     selectedTab,
     gridApiRef,
     inputSearch,
-    numberOfPatientsDisplayed: rowsProps.length,
+    patientsDisplayedCount: rowsProps.length,
     patientToRemoveForHcp,
     patientToRemoveForCaregiver,
     rowsProps,